Feature/create banner api

+++ CollabClothing.Application/Catalog/Categories/CategoryService.cs
@@ -23,7 +23,8 @@ namespace CollabClothing.Application.Catalog.Categories
     {
         private readonly CollabClothingDBContext _context;
         private readonly IStorageService _storageService;
-        private readonly string USER_CONTENT_FOLDER_NAME = "user-content/icon-category";
+        private readonly string USER_CONTENT_FOLDER_NAME = "user-content";
+        private readonly string CHILD_PATH_FOLDER_NAME = "icon-category";
 
         public CategoryService(CollabClothingDBContext context, IStorageService storageService)
         {
@@ -35,8 +36,8 @@ namespace CollabClothing.Application.Catalog.Categories
         {
             var originalFileName = ContentDispositionHeaderValue.Parse(file.ContentDisposition).FileName.Trim('"');
             var fileName = $"{Guid.NewGuid()}{Path.GetExtension(originalFileName)}";
-            await _storageService.SaveFileAsync(file.OpenReadStream(), fileName);
-            return "/" + USER_CONTENT_FOLDER_NAME + "/" + fileName;
+            await _storageService.SaveFileAsync(file.OpenReadStream(), fileName, CHILD_PATH_FOLDER_NAME);
+            return "/" + USER_CONTENT_FOLDER_NAME + "/" + CHILD_PATH_FOLDER_NAME + "/" + fileName;
         }
         public async Task<string> Create(CategoryCreateRequest request)
         {
@@ -69,6 +70,7 @@ namespace CollabClothing.Application.Catalog.Categories
         {
 
             var query = from cate in _context.Categories select new { cate };
+            var query2 = _context.Categories.Where(x => x.ParentId == x.Id).Select(o => o.NameCategory).ToString();
             if (!string.IsNullOrEmpty(request.Keyword))
             {
                 query = query.Where(x => x.cate.NameCategory.Contains(request.Keyword));
@@ -84,8 +86,10 @@ namespace CollabClothing.Application.Catalog.Categories
                              IsShowWeb = x.cate.IsShowWeb,
                              Level = x.cate.Level,
                              ParentId = x.cate.ParentId,
-                             Slug = x.cate.Slug
+                             Slug = x.cate.Slug,
+                             ParentName = query2
                          }).ToListAsync();
+
             var pageResult = new PageResult<CategoryViewModel>()
             {
                 Items = data,
@@ -97,10 +101,8 @@ namespace CollabClothing.Application.Catalog.Categories
         }
         #endregion
 
-        public async Task<ResultApi<CategoryViewModel>> GetCateById(string Id)
+        public async Task<CategoryViewModel> GetCateById(string Id)
         {
-            Category category = new Category();
-            category.CateMapping(new CategoryDTO());
             var cate = await _context.Categories.FindAsync(Id);
             if (cate == null)
             {
@@ -116,7 +118,10 @@ namespace CollabClothing.Application.Catalog.Categories
                 Level = cate.Level,
                 Slug = cate.Slug
             };
-            return new ResultApiSuccessed<CategoryViewModel>(cate1);
+
+            Category category = new Category();
+            category.CateMapping(new CategoryDTO());
+            return cate1;
         }
 
         public async Task<ResultApi<bool>> Edit(string cateId, CategoryEditRequest request)
@@ -131,21 +136,40 @@ namespace CollabClothing.Application.Catalog.Categories
             cate.ParentId = request.ParentId;
             cate.Slug = request.Slug;
             cate.Level = request.Level;
-            if (request.Icon != null)
+
+            if (cate.Icon == null)
             {
-                var fullPath = "wwwroot" + cate.Icon;
-                if (File.Exists(fullPath))
+                if (request.Icon != null)
                 {
-                    await Task.Run(() =>
-                    {
-                        File.Delete(fullPath);
-                    });
+                    cate.Icon = await this.SaveFile(request.Icon);
                 }
-                // await _storageService.DeleteFileAsync(cate.Icon);
-                cate.Icon = await this.SaveFile(request.Icon);
+                _context.Categories.Add(cate);
                 await _context.SaveChangesAsync();
+                return new ResultApiSuccessed<bool>();
             }
-            return new ResultApiSuccessed<bool>();
+            else
+            {
+                if (request.Icon != null)
+                {
+                    var fullPath = "wwwroot" + cate.Icon;
+                    if (File.Exists(fullPath))
+                    {
+                        await Task.Run(() =>
+                        {
+                            File.Delete(fullPath);
+                        });
+                    }
+                    cate.Icon = await this.SaveFile(request.Icon);
+                    _context.Categories.Update(cate);
+                    await _context.SaveChangesAsync();
+                }
+                else
+                {
+                    _context.Categories.Update(cate);
+                    await _context.SaveChangesAsync();
+                }
+                return new ResultApiSuccessed<bool>();
+            }
         }
 
         public async Task<ResultApi<bool>> Delete(string CateId)
@@ -155,10 +179,29 @@ namespace CollabClothing.Application.Catalog.Categories
             {
                 return new ResultApiError<bool>("Không tìm thấy danh mục sản phẩm");
             }
-            await _storageService.DeleteFileAsync(cate.Icon);
+
+            var fullPath = "wwwroot" + cate.Icon;
+            if (File.Exists(fullPath))
+            {
+                await Task.Run(() =>
+                {
+                    File.Delete(fullPath);
+                });
+            }
+            //await _storageService.DeleteFileAsync(cate.Icon);
             _context.Categories.Remove(cate);
             await _context.SaveChangesAsync();
             return new ResultApiSuccessed<bool>();
+        }
+
+        public async Task<List<CategoryViewModel>> GetAll()
+        {
+            var query = from c in _context.Categories select new { c };
+            return await query.Select(x => new CategoryViewModel()
+            {
+                CategoryId = x.c.Id,
+                CategoryName = x.c.NameCategory
+            }).ToListAsync();
         }
     }
 }
